Adding Data Table to better support a consistent Spreadsheet to Word/Pdf layout
Current Excel and Word layouts can be so different the common model is kind of pointless. The common model is supposed to let you build a document and not care too much about how you draw it and it will come out consistently each time.

Adding a Table and Data Table is aimed to help this.
Data Table will simply be a table of rows and columns with no merging of cells, just a straight table. This will be drawn the same on excel and word.

A Table will allow for a complex table structure to be achieved using formatting.
This will let tables to be achieved in Excel and Word.